# docs/swagger.yaml
definitions:
  gorm.DeletedAt:
    properties:
      time:
        type: string
      valid:
        description: Valid is true if Time is not NULL
        type: boolean
    type: object
  internal_coffee.Coffee:
    properties:
      createdAt:
        type: string
      deletedAt:
        $ref: '#/definitions/gorm.DeletedAt'
      description:
        example: Strong Italian coffee
        type: string
      dollar:
        example: 1
        type: number
      flag_icon:
        example: italy.png
        type: string
      id:
        type: integer
      image:
        example: espresso.jpg
        type: string
      name:
        example: Espresso
        type: string
      price:
        example: 4.99
        type: number
      ruble:
        example: 89.5
        type: number
      slug:
        example: espresso
        type: string
      updatedAt:
        type: string
    type: object
host: localhost:8081
info:
  contact: {}
  title: People-Base API documentation
  version: 1.0.1
paths:
  /coffee/create:
    post:
      consumes:
      - application/json
      description: Create coffee
      parameters:
      - description: Название кофе
        in: formData
        name: name
        required: true
        type: string
      - description: URL-friendly идентификатор
        in: formData
        name: slug
        required: true
        type: string
      - description: Цена кофе
        in: formData
        name: price
        required: true
        type: number
      - description: Описание кофе
        in: formData
        name: description
        required: true
        type: string
      - description: Цена в долларах
        in: formData
        name: dollar
        required: true
        type: number
      - description: Цена в рублях
        in: formData
        name: ruble
        required: true
        type: number
      - description: Изображение кофе
        in: formData
        name: image
        required: true
        type: file
      - description: Иконка флага страны происхождения
        in: formData
        name: flagIcon
        required: true
        type: file
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/internal_coffee.Coffee'
      summary: Create Coffee
      tags:
      - Coffee CRUD
swagger: "2.0"